requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
pydantic==2.6.4
python-dotenv==1.0.1
google-generativeai==0.5.4
slowapi==0.1.8
limits==3.7.0